% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{powstat_get_server}
\alias{powstat_get_server}
\title{Get Current PowStat Server URL}
\usage{
powstat_get_server(masked = TRUE)
}
\arguments{
\item{masked}{Logical. Whether to mask the URL in output. Default TRUE.}
}
\value{
Character. The current server URL.
}
\description{
Get Current PowStat Server URL
}
