% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{powstat_health_check}
\alias{powstat_health_check}
\title{Health Check for PowStat Server}
\usage{
powstat_health_check()
}
\value{
Logical. TRUE if the server is reachable and healthy.
}
\description{
Health Check for PowStat Server
}
