% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{powstat_set_api_token}
\alias{powstat_set_api_token}
\title{Set PowStat API Token}
\usage{
powstat_set_api_token(token)
}
\arguments{
\item{token}{Character. API token issued by the PowStat service.}
}
\value{
Invisibly returns TRUE.
}
\description{
Set PowStat API Token
}
\examples{
\dontrun{
powstat_set_api_token("your-token")
}
}
